refactor: move hooks folder to src folder
